+++ feature/login/src/main/AndroidManifest.xml
@@ -1,20 +1,3 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
-    <application
-        android:name=".TestApplication"
-        android:allowBackup="true"
-        android:label="test"
-        tools:replace="android:label">
-        <activity
-            android:name="TestActivity"
-            android:theme="@style/Theme.Material3.Light.NoActionBar"
-            android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-    </application>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android">
 </manifest>
