--- feature/login/src/main/java/com/umc/login/TestApplication.kt
@@ -1,6 +1,0 @@
-package com.umc.login
-
-import android.app.Application
-
-class TestApplication : Application() {
-}
