+++ test/login/src/main/java/com/umc/ttatta/test/login/TestApplication.kt
@@ -1,0 +1,7 @@
+package com.umc.ttatta.test.login
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class TestApplication : Application()
